refactor: split up token source into many files
